+++ internal/provider/types/backend_refs.go
@@ -10,7 +10,7 @@ type BackendRefs struct {
 	Mesh        types.String            `tfsdk:"mesh"`
 	Name        types.String            `tfsdk:"name"`
 	Namespace   types.String            `tfsdk:"namespace"`
-	Port        types.Int64             `tfsdk:"port"`
+	Port        types.Int32             `tfsdk:"port"`
 	ProxyTypes  []types.String          `tfsdk:"proxy_types"`
 	SectionName types.String            `tfsdk:"section_name"`
 	Tags        map[string]types.String `tfsdk:"tags"`
